fixed and modified the code after comments and added the color palette needed for the Start component
