+++ quiz/web/src/modules/Start/index.ts
@@ -1,0 +1,3 @@
+import Start from "modules/Start/Start";
+
+export default Start;
