--- quiz/web/src/modules/common/components/Start/index.ts
@@ -1,3 +1,0 @@
-import Start from "modules/common/components/Start/Start";
-
-export default Start;
