+++ quiz/web/src/modules/common/components/index.ts
@@ -1,5 +1,4 @@
 import ErrorBoundary from "modules/common/components/ErrorBoundary";
 import PlayerRequired from "modules/common/components/PlayerRequired";
-import Start from "modules/common/components/Start";
 
-export { ErrorBoundary, PlayerRequired, Start };
+export { ErrorBoundary, PlayerRequired };
