+++ quiz/web/tailwind.config.cjs
@@ -2,15 +2,19 @@
 
 module.exports = {
     content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+    important: "#root",
     theme: {
-        colors: {
-            primary: "#4459a9",
+        extend: {
+            container: {
+                center: true,
+            },
+            colors: {
+                primary40: "#4459a9",
+                primary70: "#92A7FD",
+                primary90: "#DDE1FF",
+                secondary90: "#DBE1FF",
+            },
         },
-        extend: {},
     },
-    plugins: [
-        {
-            tailwindcss: { config: "./tailwindcss-config.js" },
-        },
-    ],
+    plugins: [],
 };
